HTML Template
Made the HTML template and conceptualise the basic logic. Its pretty spaghetti code in the sense that the form is copy pasted, i wanted to do a for loop but its html so i got stuck. any changes i make would need to be made 6 times. need to think of a better way to fix this.